pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cognitive-core-engine"
version = "0.1.0"
description = "Core engine with API, CLI, plugins, tests and CI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.7",
    "typing-extensions>=4.8",
    "numpy>=1.26",
    "scipy>=1.11",
    "networkx>=3.2",
    "fastapi>=0.111",
    "uvicorn>=0.29",
    "httpx>=0.25",
    "rich>=13.7",
    "pyyaml>=6.0.1",
    "tqdm>=4.66",
    "opentelemetry-api>=1.24",
    "opentelemetry-sdk>=1.24",
]

[project.optional-dependencies]
api = ["fastapi>=0.111", "uvicorn>=0.29", "httpx>=0.25"]
test = ["pytest>=7.4", "pytest-cov>=4.1", "httpx>=0.25", "hypothesis>=6.99", "pyyaml>=6.0.1"]
dev = ["ruff>=0.5.0", "black>=24.4", "isort>=5.13", "mypy>=1.8", "types-requests"]

[project.scripts]
cogctl = "cognitive_core.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310","py311","py312"]

[tool.isort]
profile = "black"
line_length = 100
combine_as_imports = true
known_first_party = ["cognitive_core"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E","F","I"]
ignore = ["E203","E266","E501","F401","F841"]
src = ["src","."]

[tool.pytest.ini_options]
addopts = "-q -ra --strict-markers --disable-warnings --maxfail=1"
testpaths = ["tests"]
markers = ["integration: hits real app"]
